// apps/web/src/lib/api/generated/models/courseMonthlyAttendance.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Formex
 * OpenAPI spec version: 0.0.1
 */
import type { CourseMonthlyAttendanceStudent } from "./courseMonthlyAttendanceStudent";
import type { Program } from "./program";
import type { Subject } from "./subject";

export interface CourseMonthlyAttendance {
  id: string;
  institution_id: string;
  program_id: string;
  course_id: string;
  month: number;
  academic_year: number;
  total_classes_held: number;
  minimum_attendance_percent: number;
  /** @nullable */
  created_at: string | null;
  /** @nullable */
  updated_at: string | null;
  program?: Program;
  course?: Subject;
  attendance_students?: CourseMonthlyAttendanceStudent[];
}